+++ client/src/main/java/replicant/messages/EntityChangePayload.java
@@ -7,15 +7,18 @@ import jsinterop.base.Js;
 import org.jspecify.annotations.NonNull;
 
 /**
- * A simple abstraction for getting data off a javascript object.
+ * The serialized Entity attribute values carried by an update {@link EntityChange}.
+ *
+ * <p>This payload is not the complete Entity Change. Entity identity, target Subscription Dataset Addresses, and
+ * update or removal semantics belong to the enclosing Entity Change.</p>
  */
 @JsType(isNative = true, name = "Object", namespace = JsPackage.GLOBAL)
-public interface EntityChangeData {
+public interface EntityChangePayload {
     /**
-     * Return true if data for the attribute identified by the key is present in the change.
+     * Return true if a serialized value for the attribute identified by the key is present in the payload.
      *
      * @param key the attribute key.
-     * @return true if the data is present.
+     * @return true if the serialized value is present.
      */
     @JsOverlay
     default boolean containsKey(@NonNull final String key) {
@@ -23,10 +26,10 @@ public interface EntityChangeData {
     }
 
     /**
-     * Return true if data for the attribute identified by the key is null.
+     * Return true if the serialized value for the attribute identified by the key is null.
      *
      * @param key the attribute key.
-     * @return true if the data is null.
+     * @return true if the serialized value is null.
      */
     @JsOverlay
     default boolean isNull(@NonNull final String key) {
@@ -34,22 +37,34 @@ public interface EntityChangeData {
     }
 
     /**
-     * Return the attribute true if data for the attribute identified by the key is null.
+     * Return the integer value for the attribute identified by the key.
      *
      * @param key the attribute key.
-     * @return true if the data is null.
+     * @return the serialized integer value.
      */
     @JsOverlay
     default int getIntegerValue(@NonNull final String key) {
         return Js.asPropertyMap(this).getAsAny(key).asInt();
     }
 
+    /**
+     * Return the string value for the attribute identified by the key.
+     *
+     * @param key the attribute key.
+     * @return the serialized string value.
+     */
     @NonNull
     @JsOverlay
     default String getStringValue(@NonNull final String key) {
         return Js.asPropertyMap(this).getAsAny(key).asString();
     }
 
+    /**
+     * Return the boolean value for the attribute identified by the key.
+     *
+     * @param key the attribute key.
+     * @return the serialized boolean value.
+     */
     @JsOverlay
     default boolean getBooleanValue(@NonNull final String key) {
         return Js.asPropertyMap(this).getAsAny(key).asBoolean();